movie tableView
- switched scenes (renamed Primary to Player)
          :now Library scene is main

- can fetch categories inside category menu

TODO:

-refresh category menu after changes (category deleted, added, renamed)

-bind selected category with selected movie when checked.

-unbind...

build personal rating for menu.
