+++ src/chilltv/ChillTV.java
@@ -11,10 +11,8 @@ public class ChillTV extends Application {
     
     @Override
     public void start(Stage stage) throws Exception {
-        Parent root = FXMLLoader.load(getClass().getResource("gui/view/PrimaryScene.fxml"));
-        
+        Parent root = FXMLLoader.load(getClass().getResource("gui/view/LibraryScene.fxml"));
         Scene scene = new Scene(root);
-        
         stage.setScene(scene);
         stage.show();
