+++ src/chilltv/gui/view/PlayerScene.fxml
@@ -14,7 +14,7 @@
 <?import javafx.scene.media.MediaView?>
 <?import javafx.scene.text.Font?>
 
-<BorderPane minHeight="-Infinity" minWidth="-Infinity" prefHeight="395.0" prefWidth="640.0" xmlns="http://javafx.com/javafx/8.0.171" xmlns:fx="http://javafx.com/fxml/1" fx:controller="chilltv.gui.controller.PrimaryController">
+<BorderPane minHeight="-Infinity" minWidth="-Infinity" prefHeight="395.0" prefWidth="640.0" xmlns="http://javafx.com/javafx/8.0.171" xmlns:fx="http://javafx.com/fxml/1" fx:controller="chilltv.gui.controller.PlayerController">
    <bottom>
       <VBox fx:id="buttonBar" accessibleRole="NODE" alignment="CENTER" blendMode="SRC_ATOP" minHeight="-Infinity" minWidth="-Infinity" onMouseExited="#handle_hideBar" onMouseMoved="#handle_showBar" prefHeight="35.0" prefWidth="640.0" BorderPane.alignment="BOTTOM_CENTER">
          <children>
@@ -152,7 +152,7 @@
       </VBox>
    </bottom>
    <center>
-      <StackPane minHeight="-Infinity" minWidth="-Infinity" prefHeight="360.0" prefWidth="640.0" BorderPane.alignment="TOP_LEFT">
+      <StackPane alignment="TOP_LEFT" minHeight="-Infinity" minWidth="-Infinity" prefHeight="360.0" prefWidth="640.0" BorderPane.alignment="TOP_LEFT">
          <children>
             <MediaView fx:id="mediaView" fitHeight="360.0" fitWidth="640.0" StackPane.alignment="CENTER">
                <StackPane.margin>
